requirements.txt
pandas
seaborn
matplotlib
scikit-learn
ipykernel
streamlit
python-dotenv
sqlalchemy
psycopg2-binary
